app/backend/requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Model Inference
onnxruntime>=1.16.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0

# Model Download
requests>=2.31.0
tqdm>=4.66.0
huggingface-hub>=0.19.0

# Monitoring (Grafana Cloud)
prometheus-fastapi-instrumentator>=6.1.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For TestClient

# Optional: Development
python-dotenv>=1.0.0  # For .env file support